SSCS-3375 Create postcodeChecker and use in Contact details and postcode check
