+++ steps/start/postcode-checker/BranchForEnglandOrWales.js
@@ -1,13 +1,7 @@
 const { branch, goTo } = require('@hmcts/one-per-page');
 const { redirectTo } = require('@hmcts/one-per-page/flow');
 const { Logger } = require('@hmcts/nodejs-logging');
-const config = require('config');
-const HttpStatus = require('http-status-codes');
-const request = require('superagent');
-
-const postcodeCountryLookupUrl = config.get('postcodeChecker.url');
-const postcodeCountryLookupRegionCentres = ['glasgow'];
-const northernIslandPostcodeStart = 'bt';
+const postcodeChecker = require('utils/postcodeChecker');
 
 const logger = Logger.getLogger('PostcodeChecker.js');
 
@@ -20,27 +14,7 @@ class BranchForEnglandOrWales {
   }
 
   isEnglandOrWalesPostcode() {
-    const postcode = this.postcode;
-    if (postcode.toLocaleLowerCase().startsWith(northernIslandPostcodeStart)) {
-      return Promise.resolve(false);
-    }
-
-    return new Promise((resolve, reject) => {
-      request.get(`${postcodeCountryLookupUrl}/${postcode}`)
-        .ok(res => res.status < HttpStatus.INTERNAL_SERVER_ERROR)
-        .then(resp => {
-          if (resp.status !== HttpStatus.OK) {
-            resolve(false);
-            return;
-          }
-
-          const regionalCentre = resp.body.regionalcentre.toLocaleLowerCase();
-          resolve(!postcodeCountryLookupRegionCentres.includes(regionalCentre));
-        })
-        .catch(error => {
-          reject(error);
-        });
-    });
+    return postcodeChecker(this.postcode);
   }
 
   redirect(req, res) {
