+++ test/unit/steps/start/BranchForEnglandOrWales.test.js
@@ -1,13 +1,10 @@
 const { expect } = require('test/util/chai');
-const { merge } = require('lodash');
 const paths = require('paths');
 const proxyquire = require('proxyquire');
-const HttpStatus = require('http-status-codes');
 const sinon = require('sinon');
 
 describe('BranchForEnglandOrWales.js', () => {
   describe('Is England or Wales Postcode', () => {
-    const requireStub = {};
     const branchRedirectStub = sinon.stub();
     const redirectStub = sinon.stub();
     let branchForEnglandOrWales = null;
@@ -15,7 +12,6 @@ describe('BranchForEnglandOrWales.js', () => {
     beforeEach(() => {
       /* eslint-disable max-len */
       const BranchForEnglandOrWales = proxyquire('steps/start/postcode-checker/BranchForEnglandOrWales', {
-        superagent: requireStub,
         '@hmcts/one-per-page': {
           branch: () => {
             return { redirect: branchRedirectStub };
@@ -29,95 +25,6 @@ describe('BranchForEnglandOrWales.js', () => {
       });
       /* eslint-disable max-len */
       branchForEnglandOrWales = new BranchForEnglandOrWales('somePostcode', paths.identity.areYouAnAppointee, paths.start.invalidPostcode, paths.errors.internalServerError);
-    });
-
-    function setResponse(response) {
-      merge(requireStub, {
-        get: () => requireStub,
-        ok: () => requireStub,
-        then: handleResponse => {
-          handleResponse(response);
-          return requireStub;
-        },
-        catch: () => requireStub
-      });
-    }
-
-    function setRegionalCenterTo(regionalCentre) {
-      setResponse({ status: HttpStatus.OK, body: { regionalcentre: regionalCentre } });
-    }
-
-    it('postcode is in England', () => {
-      setRegionalCenterTo('London');
-
-      return branchForEnglandOrWales.isEnglandOrWalesPostcode()
-        .then(isEnglandOrWalesPostcode => {
-          expect(isEnglandOrWalesPostcode).to.equal(true);
-        }).catch(error => {
-          expect.fail(error);
-        });
-    });
-
-    it('postcode is in Wales', () => {
-      setRegionalCenterTo('Cardiff');
-
-      return branchForEnglandOrWales.isEnglandOrWalesPostcode()
-        .then(isEnglandOrWalesPostcode => {
-          expect(isEnglandOrWalesPostcode).to.equal(true);
-        }).catch(error => {
-          expect.fail(error);
-        });
-    });
-
-    it('postcode is in Scotland', () => {
-      setRegionalCenterTo('Glasgow');
-
-      return branchForEnglandOrWales.isEnglandOrWalesPostcode()
-        .then(isEnglandOrWalesPostcode => {
-          expect(isEnglandOrWalesPostcode).to.equal(false);
-        }).catch(error => {
-          expect.fail(error);
-        });
-    });
-
-    it('postcode starts with BT so is in Northern Island', () => {
-      branchForEnglandOrWales.postcode = 'BT1 2AB';
-
-      return branchForEnglandOrWales.isEnglandOrWalesPostcode()
-        .then(isEnglandOrWalesPostcode => {
-          expect(isEnglandOrWalesPostcode).to.equal(false);
-        }).catch(error => {
-          expect.fail(error);
-        });
-    });
-
-    it('postcode is not found', () => {
-      setResponse({ status: 404 });
-
-      return branchForEnglandOrWales.isEnglandOrWalesPostcode()
-        .then(isEnglandOrWalesPostcode => {
-          expect(isEnglandOrWalesPostcode).to.equal(false);
-        }).catch(error => {
-          expect.fail(error);
-        });
-    });
-
-    it('error getting country', () => {
-      const expectedError = 'Some error';
-
-      merge(requireStub, {
-        get: () => requireStub,
-        set: () => requireStub,
-        then: () => requireStub,
-        catch: handleError => {
-          handleError(expectedError);
-        }
-      });
-
-      return branchForEnglandOrWales.isEnglandOrWalesPostcode()
-        .catch(error => {
-          expect(error).to.equal(expectedError);
-        });
     });
 
     it('next step is /are-you-an-appointee', () => {
